S57: fix GDAL 2.1.1 regression where open options are ignored if OGR_S57_OPTIONS config option is not defined (fixes #6609)
git-svn-id: https://svn.osgeo.org/gdal/branches/2.1@34800 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/ogr/ogr_s57.py
@@ -285,6 +285,16 @@ def ogr_s57_9():
 
     gdaltest.s57_ds = None
 
+    ds = gdal.OpenEx( 'tmp/ogr_s57_9.000', open_options = ['RETURN_PRIMITIVES=ON'] )
+    if ds is None:
+        gdaltest.post_reason('fail')
+        return 'fail'
+
+    if ds.GetLayerByName('IsolatedNode') is None:
+        gdaltest.post_reason('fail')
+        return 'fail'
+    ds = None
+
     try:
         os.unlink('tmp/ogr_s57_9.000')
     except:
